requirements.txt
Django
environs
psycopg2
python-dotenv
djangorestframework
djangorestframework-simplejwt
